Extract extra output variable helper
